+++ e2e/Dockerfile
@@ -6,7 +6,7 @@ ADD ./cypress ./cypress
 ADD ./common/cypress ./common/cypress
 ADD ./common/cypress/config/e2e.json ./cypress.json
 ADD ./package.json .
-ADD ./common/src/types ./common/src/types
+ADD ./common/types ./common/types
 ADD ./dist ./dist
 
 RUN yarn global add @quasar/cli \
