+++ src/store/user/actions.js
@@ -1,6 +1,6 @@
 import { apolloClient } from '../../boot/apollo'
 import * as cookie from '../../../common/src/store/cookie'
-import types from '../../../common/src/types'
+import types from '../../../common/types'
 
 import LogIn from '../../../graphql/login.graphql'
 import SignUp from '../../../graphql/signup.graphql'
